feat: add output validation
