+++ input.txt
