--- my_ciphering_cli/AtbashStream.js
@@ -1,14 +1,0 @@
-const {Transform} = require('stream');
-const atbashEncription = require('./atbashEncription');
-
-class AtbashStream extends Transform {
-  
-  _transform(chunk, encoding, callback) {
-    let plaintext = chunk.toString().split('');
-    let ciphertext = atbashEncription(plaintext);
-    this.push(ciphertext)
-    callback();
-  }
-}
-
-module.exports = AtbashStream;
